# public/swagger-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/mail": {
      "post": {
        "operationId": "MailSenderController_create",
        "summary": "Create a new credentials",
        "description": "Create a new credentials if mandatory fields are present",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMailCredDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Credentials created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "Credentials saved successfully"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "MailSenderController_get",
        "summary": "check if creds are provided",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Credentials found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "Credentials found"
                }
              }
            }
          },
          "404": {
            "description": "Creds not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "MailSenderController_delete",
        "summary": "delete creds",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Credentials deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "Credentials deleted"
                }
              }
            }
          },
          "404": {
            "description": "Creds not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/mail-blueprint": {
      "get": {
        "operationId": "MailBlueprintCurdController_findAll",
        "summary": "Get all Mail Blueprints",
        "parameters": [
          {
            "required": false,
            "name": "query",
            "in": "query",
            "schema": {
              "example": {
                "query": {
                  "select": {
                    "name": true
                  },
                  "where": {
                    "email": "hniqbal01@gmail.com"
                  },
                  "page": 1,
                  "limit": 10
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mail Blueprints retrived successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailBlueprintPaginationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail-blueprint"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "MailBlueprintCurdController_create",
        "summary": "Create a new mailBlueprint",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMailBlueprintDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mail Blueprint created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailBlueprint"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail-blueprint"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/mail-automation": {
      "get": {
        "operationId": "MailAutomationCurdController_findAll",
        "summary": "Get all mail automations",
        "parameters": [
          {
            "required": false,
            "name": "query",
            "in": "query",
            "schema": {
              "example": {
                "query": {
                  "select": {
                    "name": true
                  },
                  "where": {
                    "email": "hniqbal01@gmail.com"
                  },
                  "page": 1,
                  "limit": 10
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mail Automation retrived successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailAutomationPaginationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail-automation",
          "v1/mail-blueprint"
        ]
      },
      "post": {
        "operationId": "MailAutomationCurdController_create",
        "summary": "Create a new mailAutomation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMailAutomationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mail Automation created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailAutomation"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail-automation",
          "v1/mail-blueprint"
        ]
      }
    },
...
    "/api/mail-blueprint/add-automation-blueprint": {
      "post": {
        "operationId": "MailBlueprintGeneralController_addMailAutomationWithTemplate",
        "summary": "Add a new blueprint and add it to automation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMailBlueprintAutomationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mail Blueprint and Automation created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailAutomation"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail-blueprint"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/mail-blueprint/add-automation": {
      "post": {
        "operationId": "MailBlueprintGeneralController_AddMailAutomation",
        "summary": "Add or update an automation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMailAutomationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mail Automation created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailAutomation"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail-blueprint"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/mail-blueprint/send-blueprint": {
      "post": {
        "operationId": "MailBlueprintGeneralController_sendBluePrintMail",
        "summary": "Send email by blueprint",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMailBlueprintDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Send email by blueprint",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "done"
                }
              }
            }
          },
          "400": {
            "description": "Missing placeholders",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail-blueprint"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Boilerplate v1 API",
    "description": "Basic docs for API in version 1",
    "version": "1.0",
    "contact": {
      "name": "Shirsho ",
      "url": "https://marketron.com",
      "email": "hniqbal01@gmail.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ExceptionResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "example": "Invalid request parameters"
          },
          "timestamp": {
            "type": "string",
            "example": "2024-11-22T04:10:35.729Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/resource"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "timestamp",
          "path"
        ]
      },
      "CreateMailCredDTO": {
        "type": "object",
        "properties": {
          "smtpHost": {
            "type": "string",
            "example": "asd"
          },
          "smtpPort": {
            "type": "string",
            "example": "asd"
          },
          "smtpFrom": {
            "type": "string",
            "example": "asd"
          },
          "smtpUser": {
            "type": "string",
            "example": "asd"
          },
          "smtpPassword": {
            "type": "string",
            "example": "asd"
          }
        },
        "required": [
          "smtpHost",
          "smtpPort",
          "smtpFrom",
          "smtpUser",
          "smtpPassword"
        ]
      },
...
      "MailBlueprint": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The unique name of the mail blueprint",
            "example": "WelcomeEmail"
          },
          "bodyContent": {
            "type": "string",
            "description": "The content/body of the email",
            "example": "Hello, welcome to our platform!"
          },
          "subjectContent": {
            "type": "string",
            "description": "The subject line of the email",
            "example": "Welcome to Our Platform"
          },
          "placeholders": {
            "description": "List of placeholder keys that can be replaced in the content",
            "example": [
              "username",
              "date"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "bodyContent",
          "subjectContent",
          "placeholders"
        ]
      },
      "MailBlueprintPaginationResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MailBlueprint"
            }
          },
          "count": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "data",
          "count"
        ]
      },
      "CreateMailBlueprintDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Blueprint name"
          },
          "bodyContent": {
            "type": "string",
            "example": "Shirsho {{placeholder}}"
          },
          "subjectContent": {
            "type": "string",
            "example": "Shirsho {{placeholder}}"
          }
        },
        "required": [
          "name",
          "bodyContent",
          "subjectContent"
        ]
      },
...
      "MailAutomation": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the mail automation",
            "example": "WelcomeEmail"
          },
          "resourceName": {
            "type": "string",
            "description": "The resource name of the mail automation",
            "example": "exam"
          },
          "bluePrintId": {
            "type": "string",
            "description": "The blueprint id of the mail automation",
            "example": "mongo id"
          }
        },
        "required": [
          "resourceName",
          "bluePrintId"
        ]
      },
      "MailAutomationPaginationResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MailAutomation"
            }
          },
          "count": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "data",
          "count"
        ]
      },
      "CreateMailAutomationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Shirsho",
            "description": "Name of the mail automation optional"
          },
          "resourceName": {
            "type": "string",
            "example": "example"
          },
          "bluePrintId": {
            "type": "string",
            "example": "mongo id"
          }
        },
        "required": [
          "resourceName",
          "bluePrintId"
        ]
      },
...
      "CreateMailBlueprintAutomationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Blueprint name"
          },
          "resourceName": {
            "type": "string",
            "example": "course"
          },
          "bodyContent": {
            "type": "string",
            "example": "Shirsho {{placeholder}}"
          },
          "subjectContent": {
            "type": "string",
            "example": "Shirsho {{placeholder}}"
          }
        },
        "required": [
          "name",
          "resourceName",
          "bodyContent",
          "subjectContent"
        ]
      },
      "SendMailBlueprintDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "mongo id"
          },
          "to": {
            "type": "string",
            "example": "hniqbal01@gmail.com"
          },
          "cc": {
            "example": [
              "hniqbal01@gmail.com",
              "hniqbal01@gmail.com"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "bcc": {
            "example": [
              "hniqbal01@gmail.com",
              "hniqbal01@gmail.com"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "placeData": {
            "type": "object",
            "description": "JSON object with placeholders and their values. can be a nested json",
            "example": {
              "adifferent_placeholder": "Shirsho",
              "placeholderwithspace": "<h1>ExampleCorp</h1>"
            }
          }
        },
        "required": [
          "id",
          "to",
          "placeData"
        ]
      }
    }
  }
}